requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.13
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
pymysql>=1.0.3
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
alembic>=1.10.4
python-dotenv>=1.0.0
bcrypt>=4.0.1
httpx>=0.24.0
Pillow>=9.0.0

pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
factory-boy>=3.2.1
faker>=18.0.0